requirements.txt
attrs==22.2.0
bandit==1.7.4
gitdb==4.0.10
GitPython==3.1.30
iniconfig==2.0.0
mypy==1.0.0
mypy-extensions==1.0.0
packaging==23.0
pbr==5.11.1
pluggy==1.0.0
pydocstyle==6.3.0
pytest==7.2.1
python-dotenv==0.21.1
PyYAML==6.0
smmap==5.0.0
snowballstemmer==2.2.0
stevedore==5.0.0
typing_extensions==4.5.0
